+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrelease]
+
+### Added
+- Add code editor to reportsV2
+
 ## Fixed
 - Fixed dashDTPicker active element text color
 
@@ -16,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created gantt class
 - Created piechart class
 - Created graph class
-- Adding a code editor to data exploration
 
 ### Fixed
 - Multiline. Some chart line display settings do not work