+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.91.0]
+
+### Added
+- Add code editor to reportsV2
+- add visualisation dashAccumulators.vue
+
+
+### Fixed
+- Fixed dashDTPicker active element text color
+- Fixed tokens in csvg visualisation
+- Fixed csvg popup
+- Fixed dashDTPicker active element text color
+- Fixed color picker background
+
 ## [1.90.0]
 
 ### Added